MISC: fix yarn corepack
fixes: Unable to install local packages for development
Fixes #5

turns out I had a local override in my local environment that was setting a later node version so didn't realise it was broken for other people. This changes to the recommended way of installing yarn with corepack.

it also bumps node to v20 which is required for corepack. I think its very unlikley this module will use core CMS js given how much of a departure the JS structure is anyway.